requirements.txt
# Python Requirements for PDF Processing and Language Detection

# PDF Extraction
pdfminer.six>=20221105

# Language Detection
langdetect>=1.0.9

# OCR (Optional - for scanned PDFs)
# Requires: brew install tesseract tesseract-lang (macOS)
# or: apt-get install tesseract-ocr tesseract-ocr-heb tesseract-ocr-ell (Linux)
pytesseract>=0.3.10
Pillow>=10.0.0

# Better PDF Processing (Optional)
pdfplumber>=0.10.0

# Web Scraping (for Hebrew download script)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Advanced Language Detection (Optional)
# polyglot>=16.7.4
# pyicu>=2.11
# pycld2>=0.41



